The unit-tests facility proved to be useless in a strange way - it determines errors in float operations, that in other projects are not present. And starting a test takes an enormous amount of time. The ReciveUpdate feature, where it can decline update, if it holds the same state, just in a past also dosn't work. Somehow, recived state with a lag is newer, then the one, that is currently holded by an entity. This needs further exploration

+++ GameSimulation/GameTime.h
@@ -3,8 +3,9 @@
 
 namespace GEM::GameSim
 {
-	using GameTime = uint64_t;
+	using GameTime = int64_t;
+
 
 	inline GameTime SecondsToGameTimeInterval(float t) { return static_cast<GameTime>(t * 100000); }
-	inline float GameTimeToSeconds(GameTime t) { return static_cast<float>(t) / 100000.0f; }
+	inline float GameTimeToSeconds(GameTime t) { return t / 100000.0f; }
 }
